backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
drf-nested-routers==0.94.2
django-cors-headers==4.9.0
django-safedelete==1.4.1
django-filter==25.2
django-money==3.5.3
django-environ==0.13.0
psycopg2-binary==2.9.11
Pillow==11.0.0
celery==5.4.0
redis==5.2.1
gunicorn==23.0.0
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-django==4.12.0
factory-boy==3.3.1
channels==4.3.2
channels-redis==4.2.1
daphne==4.1.2
